+++ packages/app/src/components/Admin/UserManagement.module.scss
@@ -1,5 +1,5 @@
 // styles for admin user search
 .search-clear :global {
-  top: 90px;
+  top: 7px;
   right: 4px;
 }
